requirements.txt
numpy
sentence_transformers
faiss-cpu
PyPDF2
requests
streamlit